Converted DynamicRealmTests to JUnit4. Added JUnit test guidelines.
