+++ CONTRIBUTING.md
@@ -25,15 +25,119 @@ Realm welcomes all contributions! The only requirement we have is that, like man
 
 [Please submit your CLA electronically using our Google form](https://docs.google.com/forms/d/1bVp-Wp5nmNFz9Nx-ngTmYBVWVdwTyKj4T0WtfVm0Ozs/viewform?fbzx=4154977190905366979) so we can accept your submissions. The GitHub username you file there will need to match that of your Pull Requests. If you have any questions or cannot file the CLA electronically, you can email <help@realm.io>.
 
-## Unit tests
+## Repository Guidelines
 
-All PR's must be accompanied by related unit tests.
-
-## Code style
+### Code Style
 
 While we havn't described our code style yet, please just follow the existing style you see in the files you change.
 
-## Javadoc
+### Unit Tests
+
+All PR's must be accompanied by related unit tests. All bug fixes must have a unit test proving that the bug is fixed.
+You can use `./realm/gradlew connectedCheck createDebugCoverageReport` to generate a coverage report to check for 
+missing unit test coverage. The aim is 100% code coverage.
+
+When writing unit tests, use the following guide lines:
+
+1) Unit tests must be written using JUnit4.
+
+2) All tests for a class should be grouped in a class called `<className>Tests`, unless the functionality is cross-
+   cutting like [`RxJavaTests`](https://github.com/realm/realm-java/blob/master/realm/realm-library/src/androidTest/java/io/realm/RxJavaTests.java) 
+   or [`RealmAsyncQueryTests`](https://github.com/realm/realm-java/blob/master/realm/realm-library/src/androidTest/java/io/realm/RealmAsyncQueryTests.java).
+
+3) Test methods should use camelCase and underscore `_` between logical sections to increase method name readability. 
+   Methods should ideally start with the name of the method being tested. Patterns like: `<methodName>_<description>`, 
+   `<methodName>_<param>_<description>` or `<description>` are encouraged.
+   
+4) All unit tests creating Realms must do so using the [`TestRealmConfigurationFactory`](https://github.com/realm/realm-java/blob/master/realm/realm-library/src/androidTest/java/io/realm/rule/TestRealmConfigurationFactory.java) 
+   or [`RunInLooperThread`](https://github.com/realm/realm-java/blob/master/realm/realm-library/src/androidTest/java/io/realm/rule/RunInLooperThread.java) 
+   test rules. This ensures that all Realms are properly closed and deleted between each test.
+
+5) Use the `@RunInLooperThread` rule for any test that depends on Realms notification system. 
+
+6) Input-parameters should be boundary tested. Especially `Null/NotNull`, but also the state of Realm objects like
+   standalone objects, deleted objects, objects from other threads.
+
+7) Unit tests are not required to only have 1 test. It is acceptable to combine multiple tests into one unit test, but
+   if it fails, it should be clear why it failed. E.g. you can group related tests with the same setup like negative 
+   tests. If you do so, make sure to separate each "subtest" with a comment stating what you test.
+
+8) Use only `@Test(expected = xxx.class)` or the `ExceptedException` rule to detect exceptions if it is the last
+   line of the test that is expected to throw. Otherwise use the following:
+   
+    try {
+      somethingThatThrowsIllegalArgument();   
+    } catch (IllegalArgumentException ignored) {
+    }
+
+9) Use comments to make the intent of the unit test easily understandable at a glance. A simple one line comment is 
+   often easier to read `thanALongCamelCasedSentenceThatAttemptsToDescribeWhatHappens`. Describe the test steps inside 
+   the method, if it's not glaringly obvious.
+
+This is an example of how a unit test class could look like:
+
+    @RunWith(AndroidJUnit4.class)
+    public class RealmTests {
+    
+      @Rule
+      public final TestRealmConfigurationFactory configFactory = new TestRealmConfigurationFactory();
+    
+      @Rule
+      public final RunInLooperThread looperThread = new RunInLooperThread();
+    
+      private Realm realm;   
+    
+      @Before
+      public void setUp() {
+         RealmConfiguration config = configFactory.createConfiguration();
+         realm = Realm.getInstance(config);
+      }
+      
+      @After
+      public void tearDown() {
+        if (realm != null) {
+            realm.close();
+        }  
+      }
+    
+      @Test(expected = IllegalStateException.class)
+      public void createObject_outsideTransaction() {
+        realm.createObject(Foo.class);
+      }
+    
+      @Test
+      public void createObject_illegalInput {
+        // Class not part of the schema
+        try {
+          realm.createObject(Foo.class);    
+        } catch (IllegalArgumentException ignored) {
+        }
+    
+        // Null class
+        try {
+            realm.createObject(null);    
+        } catch (IllegalArgumentException ignored) {
+        }
+      }
+      
+      @Test
+      @RunTestInLooperThread
+      public void addChangeListener_notifiedOnLocalCommit() {
+        realm.addChangeListener(new RealmChangeListener() {
+            @Override
+            public void onChange() {
+                assert(1, realm.allObjects(Foo.class).size());
+                looperThread.testComplete();
+            }
+        });
+    
+        realm.beginTransaction();
+        realm.createObject(Foo.class);
+        realm.commitTransaction();
+      }
+    }
+  
+### Javadoc
 
 All public classes and methods must have Javadoc describing their purpose.
