+++ realm/realm-library/findbugs-filter.xml
@@ -42,4 +42,12 @@
         <Field name="minDepth" />
         <Bug pattern="URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD" />
     </Match>
+
+    <!-- Unit tests -->
+    <Match>
+        <Class name="io.realm.DynamicRealmTests"/>
+        <Method name="allObjectsSorted_failures" />
+        <Bug pattern="NP_NULL_PARAM_DEREF_ALL_TARGETS_DANGEROUS"/>
+    </Match>
+
 </FindBugsFilter>
